Add wave limitation (only particles aiming for the screen are drawn). with smaller-than-window debug rect.

+++ src/Wave.cpp
@@ -1,4 +1,6 @@
 #include "Wave.h"
+#define WINDOW_W 1500
+#define WINDOW_H 1000
 
 //default constructor
 Wave::Wave(){ }
@@ -15,14 +17,43 @@ Wave::Wave(float x, float y){
     mesh.setMode(OF_PRIMITIVE_POINTS);
     
     double pitch = TWO_PI/resolution;                                       // 'angle' between each point
-    
-    
-    
-    for (int i=0; i<resolution; i++){
-        mesh.addVertex(ofPoint(x,y)); mesh.addColor(ofFloatColor(1,1,1));   // add vertexes to mesh, and create new particles
-        particles.push_back(Particle(i, x, y, pitch*i, 4));                 // FIXME: particles are a double of mesh verteces.
+
+    for (float i=0; i<TWO_PI; i+=pitch){
+        if(rayIntersects(x, y, i)){                                             // only add particles that are aimed to the screen
+            mesh.addVertex(ofPoint(x,y)); mesh.addColor(ofFloatColor(1,1,1));   // add vertexes to mesh, and create new particles
+            particles.push_back(Particle(x, y, i, 4));                          // FIXME: particles are a double of mesh verteces.
+        }
     }
+    particles[0].isEdge = true;
+    particles[particles.size()-1].isEdge = true;
 }
+
+// checks if a ray ( defined by an origin and an angle(/direction) ) crosses(intersects) the screen
+bool Wave::rayIntersects(int x, int y, float angle){
+    float a = tan(angle);
+    float b = y-(a*x);
+    int side = (cos(angle)>0)? 1 : -1;
+    
+    float px, py;
+    
+    // intersection with left side of rect
+    px = 0; py = (a*px)+b;
+    if( ( py < WINDOW_H ) && ( py > 0 ) && (x-px)*side<0) return true;
+
+    // intersection with right side of rect
+    px = WINDOW_W; py = (a*px)+b;
+    if( ( py < WINDOW_H ) && ( py > 0 ) && (x-px)*side<0) return true;
+    
+    // intersection with bottom side of rect
+    py = WINDOW_H; px = (py-b)/a;
+    if( (px>0) && (px<WINDOW_W) && (y-py)*side*a<0) return true;
+
+    // intersection with top side of rect
+    py = 0; px = (py-b)/a;
+    if( (px>0) && (px<WINDOW_W) && (y-py)*side*a<0) return true;
+//
+    return false;
+};
 
 void Wave::update(vector<PointElement *>& elements, float opacity){
     
@@ -30,22 +61,24 @@ void Wave::update(vector<PointElement *>& elements, float opacity){
     int os = elements.size();
     for (int i=0; i<s; i++) {
 
-        // black hole deletion
+        // black hole slurp
         if(blackHole != NULL){
             particles[i].position += (blackHole->pos - particles[i].position)/ofRandom(2,20);        // ease to black hole, with some random.
             if(particles[i].position.distance(blackHole->pos) < 1) particles[i].alive = false;       // arrived at blackhole, delete this particle
         }
         else{
-            particles[i].update(speed);                                              // update particle position
-            for (int j=0; j < os; j++) {
-                if(elements[j]->collisionCheck(                                     // (bounding box) collision check
-                   particles[i].position,
-                   particles[(i+1)%s].position,
-                   particles[(i+1)%s].pPosition,
-                   particles[i].pPosition
-                   )){
-                    elements[j]->collided();
-                    if(elements[j]->kind == DESTROYER_ELEMENT) blackHole = elements[j];  // DESTROYER_ELEMENT -> destroy wave on collision
+            particles[i].update(speed);                                                // update particle position
+            if(i<s-1){                                                                 // do not check edge points to avoid 'wrap' bugs.
+                for (int j=0; j < os; j++) {
+                    if(elements[j]->collisionCheck(                                    // (bounding box) collision check
+                       particles[i].position,
+                       particles[(i+1)].position,
+                       particles[(i+1)].pPosition,
+                       particles[i].pPosition
+                       )){
+                        elements[j]->collided();
+                        if(elements[j]->kind == DESTROYER_ELEMENT) blackHole = elements[j];  // DESTROYER_ELEMENT -> destroy wave on collision
+                    }
                 }
             }
         }
@@ -66,6 +99,8 @@ void Wave::update(vector<PointElement *>& elements, float opacity){
 }
 
 void Wave::draw(){
+    ofSetColor(255);
+    ofRect(0, 0, WINDOW_W, WINDOW_H);
     glPointSize(3);
     mesh.draw();
     glPointSize(1);
